refactor: Path와 Fare 통합
